Unify matrix/sparse network fabrication via StageMatrix trait
Extract the shared feedforward fabrication logic (dependency resolution,
carry propagation, output reordering) into a single generic
`fabricate_stages<N, E, M: StageMatrix>` function in `network::mod`.

Each backend implements `StageMatrix::from_stage_data` to convert the
common triplet representation into its native matrix type (`DMatrix<f64>`
or `CscMatrix<f64>`). Both fabricators become thin wrappers that call
`fabricate_stages` and wrap the result, eliminating ~200 lines of
duplicated logic.

Also removes the now-unused `type Matrix = Vec<Vec<f64>>` alias.
