+++ src/matrix/feedforward/fabricator.rs
@@ -1,17 +1,21 @@
-use crate::network::{EdgeLike, Fabricator, NetworkLike, NodeLike};
-use nalgebra::{DMatrix, DVector};
-use std::collections::HashMap;
+use crate::network::{fabricate_stages, EdgeLike, Fabricator, NetworkLike, NodeLike, StageMatrix};
+use nalgebra::DMatrix;
 
 pub struct MatrixFeedforwardFabricator;
 
-impl MatrixFeedforwardFabricator {
-    fn get_matrix(dynamic_matrix: Vec<Vec<f64>>) -> DMatrix<f64> {
-        let columns = dynamic_matrix
-            .into_iter()
-            .map(DVector::from_vec)
-            .collect::<Vec<_>>();
-
-        DMatrix::from_columns(&columns)
+impl StageMatrix for DMatrix<f64> {
+    fn from_stage_data(
+        rows: usize,
+        cols: usize,
+        row_indices: Vec<usize>,
+        col_indices: Vec<usize>,
+        data: Vec<f64>,
+    ) -> Self {
+        let mut m = DMatrix::zeros(rows, cols);
+        for ((&c, &r), &v) in col_indices.iter().zip(row_indices.iter()).zip(data.iter()) {
+            m[(r, c)] = v;
+        }
+        m
     }
 }
 
@@ -23,198 +27,10 @@ where
     type Output = super::evaluator::MatrixFeedforwardEvaluator;
 
     fn fabricate(net: &impl NetworkLike<N, E>) -> Result<Self::Output, &'static str> {
-        // build dependency graph by collecting incoming edges per node
-        let mut dependency_graph: HashMap<usize, Vec<&E>> = HashMap::new();
-
-        for edge in net.edges() {
-            dependency_graph
-                .entry(edge.end())
-                .and_modify(|dependencies| dependencies.push(edge))
-                .or_insert_with(|| vec![edge]);
-        }
-
-        if dependency_graph.is_empty() {
-            return Err("no edges present, net invalid");
-        }
-
-        // keep track of dependencies present
-        let mut dependency_count = dependency_graph.len();
-
-        // println!("initial dependency_graph {:#?}", dependency_graph);
-
-        // contains list of matrices (stages) that form the computable net
-        let mut compute_stages: Vec<crate::Matrix> = Vec::new();
-        // contains activation functions corresponding to each stage
-        let mut stage_transformations: Vec<crate::Transformations> = Vec::new();
-        // set available nodes a.k.a net input
-        let mut available_nodes = net.inputs();
-        // sort via Ord implementation of provided nodes to guarantee each input will be processed by the same node every time
-        available_nodes.sort_unstable();
-        // reduce nodes to ids
-        let mut available_nodes: Vec<usize> = available_nodes.iter().map(|n| n.id()).collect();
-
-        // println!("available_nodes {:?}", available_nodes);
-
-        // set wanted nodes a.k.a net output
-        let mut wanted_nodes = net.outputs();
-        // sort via Ord implementation of provided nodes to guarantee each output will appear in the same order every time
-        wanted_nodes.sort_unstable();
-        // reduce nodes to ids
-        let wanted_nodes: Vec<usize> = wanted_nodes.iter().map(|n| n.id()).collect();
-
-        // println!("wanted_nodes {:?}", wanted_nodes);
-
-        // gather compute stages by finding computable nodes and required carries until all dependencies are resolved
-        while !dependency_graph.is_empty() {
-            // setup new compute stage
-            let mut stage_matrix: crate::Matrix = Vec::new();
-            // setup new transformations
-            let mut transformations: crate::Transformations = Vec::new();
-            // list of nodes becoming available by compute stage
-            let mut next_available_nodes: Vec<usize> = Vec::new();
-
-            for (&dependent_node, dependencies) in dependency_graph.iter() {
-                // marker if all dependencies are available
-                let mut computable = true;
-                // eventual compute vector
-                let mut compute_or_carry = vec![f64::NAN; available_nodes.len()];
-                // check every dependency
-                for &dependency in dependencies {
-                    let mut found = false;
-                    for (index, &id) in available_nodes.iter().enumerate() {
-                        if dependency.start() == id {
-                            // add weight to compute vector at position of input
-                            compute_or_carry[index] = dependency.weight();
-                            found = true;
-                        }
-                    }
-                    // if any dependency is not found the node is not computable yet
-                    if !found {
-                        computable = false;
-                    }
-                }
-                if computable {
-                    // replace NAN with 0.0
-                    for n in &mut compute_or_carry {
-                        if n.is_nan() {
-                            *n = 0.0
-                        }
-                    }
-                    // add vec to compute stage
-                    stage_matrix.push(compute_or_carry);
-                    // add activation function to stage transformations
-                    transformations.push(
-                        net.nodes()
-                            .iter()
-                            .find(|&node| node.id() == dependent_node)
-                            .unwrap()
-                            .activation(),
-                    );
-                    // mark node as available in next iteration
-                    next_available_nodes.push(dependent_node);
-                } else {
-                    // figure out carries
-                    for (index, &weight) in compute_or_carry.iter().enumerate() {
-                        // if there is some partial dependency that is not carried yet
-                        if !next_available_nodes.contains(&available_nodes[index])
-                            && !weight.is_nan()
-                        {
-                            let mut carry = vec![0.0; available_nodes.len()];
-                            carry[index] = 1.0;
-                            // add carry vector
-                            stage_matrix.push(carry);
-                            // add identity function for carried vector
-                            transformations.push(|val| val);
-                            // add node as available
-                            next_available_nodes.push(available_nodes[index]);
-                        }
-                    }
-                }
-            }
-
-            // keep any wanted notes if available (output)
-            for wanted_node in wanted_nodes.iter() {
-                for (index, available_node) in available_nodes.iter().enumerate() {
-                    if available_node == wanted_node {
-                        // carry only if not carried already
-                        if !next_available_nodes.contains(available_node) {
-                            let mut carry = vec![0.0; available_nodes.len()];
-                            carry[index] = 1.0;
-                            // add carry vector
-                            stage_matrix.push(carry);
-                            // add identity function for carried vector
-                            transformations.push(|val| val);
-                            // add node as available
-                            next_available_nodes.push(*available_node);
-                        }
-                    }
-                }
-            }
-
-            // remove resolved dependencies from dependency graph
-            for node in next_available_nodes.iter() {
-                dependency_graph.remove(node);
-            }
-
-            // if no dependency was removed no progess was made
-            if dependency_graph.len() == dependency_count {
-                return Err("can't resolve dependencies, net invalid");
-            } else {
-                dependency_count = dependency_graph.len();
-            }
-
-            // println!("next_available_nodes {:?}", next_available_nodes);
-
-            // reorder last stage according to net output order (invalidates next_available_nodes order which wont be used after this point)
-            if dependency_graph.is_empty() {
-                // println!("stage_matrix {:?}", stage_matrix);
-
-                let mut reordered_matrix = stage_matrix.clone();
-                let mut reordered_transformations = transformations.clone();
-
-                let mut matched_wanted_count = 0;
-
-                for ((available_node, column), transformation) in next_available_nodes
-                    .iter()
-                    .zip(stage_matrix.into_iter())
-                    .zip(transformations.into_iter())
-                {
-                    for (index, wanted_node) in wanted_nodes.iter().enumerate() {
-                        if available_node == wanted_node {
-                            reordered_matrix[index] = column;
-                            reordered_transformations[index] = transformation;
-                            matched_wanted_count += 1;
-                            break;
-                        }
-                    }
-                }
-
-                if matched_wanted_count < wanted_nodes.len() {
-                    return Err(
-                        "dependencies resolved but not all outputs computable, net invalid",
-                    );
-                }
-
-                // println!("reordered_matrix {:?}", reordered_matrix);
-
-                stage_matrix = reordered_matrix;
-                transformations = reordered_transformations;
-            }
-
-            // add resolved dependencies and transformations to compute stages
-            compute_stages.push(stage_matrix);
-            stage_transformations.push(transformations);
-
-            // set available nodes for next iteration
-            available_nodes = next_available_nodes;
-        }
-
+        let (stages, transformations) = fabricate_stages::<N, E, DMatrix<f64>>(net)?;
         Ok(super::evaluator::MatrixFeedforwardEvaluator {
-            stages: compute_stages
-                .into_iter()
-                .map(MatrixFeedforwardFabricator::get_matrix)
-                .collect(),
-            transformations: stage_transformations,
+            stages,
+            transformations,
         })
     }
 }
